Using the 'pacman' technique
The pacman technique is computing the result as we go - sort of ..,
The idea is to always maintain the highest values for each position in the line.
We only need to get the max at the end.

The one trick is that we are lucky that list[-1] is the last element that's why
we append zero to the var that holds the current values at each each iteration
we can add this 0 to the first element of the current line of the triangle

Tests:
732506 is the expected output as this was the only email that I was able to send
to the company that created this exercice.