Are two strings permutation of each others
Simple methods to test if two strings are permutation of each others.

is_permutation sorts both strings and return True if the result is the
same, False otherwise. Sorts takes O(n log n)

array_permutation creates an array of 128 zeros to count the occurrence
of each ASCII char in the strings. Increment the correct position in the
array for str1 and decrement the correct position for each char of str2.
If the min and max of the array is 0 the strings are permutations.
We go through the strings once - O(n)