[LRU] Prevent cache size from going over capacity
If any key is being added that would bring the cache size beyond the
capacity the least recently used key is evicted.

We maintain an in order "access log" for keys so that the latest key
added or accessed will always be the head of that log. The least
recently access entry is last and is popped out whenever the addition of
a key would result in going over the set capacity.

Tests:
- lru tests OK