[AUDIO-IMG] First attempt at writing music
We compute the number of "wave frames" by multiplying the estimated
duration by the sample rate (known from the recording application).
We then divide that by the width of the image (or len of the min/max)
list we've computed. This gives us the number of "wave frame" for each
column of pixel.

We go over the list of min/max, we arbitrarily decide to start with one
(based on modulo of current index), scale the value for an 8-bit sample
wave file and repeat that value for the number of "wave frame" as
defined above.

Well that doesn't work :)

The output file isn't any audible ... but we go on!