[PIXELS] Create the ORANGE color
That color does not "naturally" occur and can only result out of the
combination of RED and YELLOW. It's input representation will result in
two Pixels being created at the given position (1 red, 1 yellow).

When a red and yellow pixels move to the same position it results in an
ORANGE pixel that will then "separate" back into the two original
pixels.

ORANGE could be considered a "virtual" pixel, for a lack of a better
word.

Tests:
- ./gradlew test